+++ Src/FluentAssertions/Equivalency/EquivalencyValidationContext.cs
@@ -10,7 +10,6 @@ namespace FluentAssertions.Equivalency
     /// </summary>
     public class EquivalencyValidationContext : IEquivalencyValidationContext
     {
-        private static readonly ComplexTypeMap TypeMap = new();
         private Tracer tracer;
 
         public EquivalencyValidationContext(INode root, IEquivalencyAssertionOptions options)
@@ -72,7 +71,8 @@ namespace FluentAssertions.Equivalency
 
         public bool IsCyclicReference(object expectation)
         {
-            bool isComplexType = TypeMap.IsComplexType(expectation);
+            bool isComplexType = expectation is not null && Options.GetEqualityStrategy(expectation.GetType())
+                is EqualityStrategy.Members or EqualityStrategy.ForceMembers;
 
             var reference = new ObjectReference(expectation, CurrentNode.PathAndName, isComplexType);
             return CyclicReferenceDetector.IsCyclicReference(reference, Options.CyclicReferenceHandling, Reason);