+++ Src/FluentAssertions/Extensions/FluentDateTimeExtensions.cs
@@ -135,7 +135,7 @@ namespace FluentAssertions.Extensions
         /// </summary>
         public static DateTime At(this DateTime date, TimeSpan time)
         {
-            return new DateTime(date.Year, date.Month, date.Day, time.Hours, time.Minutes, time.Seconds);
+            return date.Date + time;
         }
 
         /// <summary>
@@ -154,7 +154,7 @@ namespace FluentAssertions.Extensions
                 throw new ArgumentOutOfRangeException(nameof(nanoseconds), "Valid values are between 0 and 999");
             }
 
-            var value = new DateTime(date.Year, date.Month, date.Day, hours, minutes, seconds, milliseconds);
+            var value = new DateTime(date.Year, date.Month, date.Day, hours, minutes, seconds, milliseconds, date.Kind);
 
             if (microseconds != 0)
             {